+++ CRM/Civirules/Upgrader.php
@@ -666,6 +666,13 @@ class CRM_Civirules_Upgrader extends CRM_Civirules_Upgrader_Base {
     return true;
   }
 
+  public function upgrade_2038()
+  {
+    // Add the action: add membership
+    CRM_Civirules_Utils_Upgrader::insertActionsFromJson($this->extensionDir . DIRECTORY_SEPARATOR . 'sql/actions.json');
+    return true;
+  }
+
 
 }